% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-database.R
\name{build_sqlite}
\alias{build_sqlite}
\title{Build the SQLite database}
\usage{
build_sqlite(
  csv_folder,
  sql_folder = csv_folder,
  sql_filename = "db.sqlite",
  vroom_guess_max = 1e+05,
  overwrite = F,
  ...
)
}
\arguments{
\item{csv_folder}{the location of the CSV files provided by ISARIC. This
should contain the CSV files from a single release only, and should not
contain any CSV files other than those CSV files. Also, there should be no two
CSV files whose names starts with the same first two letters.}

\item{sql_folder}{the location to save the SQLite database. Defaults to
\code{csv_folder}.}

\item{sql_filename}{the filename to use for the SQLite database, that will
be saved in \code{sql_folder}.}

\item{vroom_guess_max}{the maximum number of rows to use for gussing data types
in \code{\link[vroom:vroom]{vroom::vroom()}} (when reading CSVs).}

\item{overwrite}{logical where \code{TRUE} causes the function to
overwrite existing tables with the same name in the SQLite database if it
exists already.}

\item{...}{optional arguments passed to \code{\link[DBI:dbWriteTable]{DBI::dbWriteTable()}}.}
}
\value{
The function returns \code{TRUE}, after creating the SQLite database
in the folder \code{sql_folder}. Returns \code{FALSE} if all tables
already exist and \code{overwrite=FALSE}.
}
\description{
This function uses the CSV files provided by ISARIC to build a SQLite database
of tables, where each table will be named after the first two letters of the
corresponding CSV file, e.g., DM, DS, LB, HO, etc. This function can take
some time to execute, e.g., 20 minutes.
}
\examples{

build_sqlite(csv_folder="C:\\\\Datasets\\\\AKI\\\\data\\\\raw\\\\07APR2021\\\\csvs")

}
