% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{process_occur}
\alias{process_occur}
\title{Categorise occurrence of events}
\usage{
process_occur(data, xxOCCUR, xxPRESP, overwrite = F)
}
\arguments{
\item{data}{a \code{data.frame} or SQL connection (\code{tbl_sql}) object}

\item{xxOCCUR}{the name of OCCUR column (e.g., SAOCCUR or HOOCCUR)}

\item{xxPRESP}{the name of the PRESP column (e.g., SAPRESP or HOPRESP)}

\item{overwrite}{a logical value. \code{TRUE} means the column \code{status} will
be overwritten if already present in \code{data}. \code{FALSE} means the function
will throw an error if \code{status} is already present in \code{data}.}
}
\value{
Returns the \code{data} argument with an additional column called \code{status},
having values 'yes', 'no', 'unknown' or 'categorisation failed'.
}
\description{
Use the OCCUR and PRESP columns from the ISARIC
\href{https://en.wikipedia.org/wiki/SDTM}{SDTM} data format to derive
a column indicating the occurrence status of each event.
}
\details{
This function implements the following logic:\preformatted{xxOCCUR == Y                 -> yes
xxOCCUR == NA & xxPRESP != Y -> yes
xxOCCUR == N                 -> no
xxOCCUR == NA & xxPRESP == Y -> unknown
xxOCCUR == U                 -> unknown
}
}
\examples{

con <- DBI::dbConnect(RSQLite::SQLite(), DIRS$db)
ho <- dplyr::tbl(con, "HO")
ho_modified <- ho \%>\%
   process_occur(xxOCCUR = HOOCCUR, xxPRESP = HOPRESP)

}
